+++ src/main/java/com/example/finca_hexagonal/domain/ports/in/horario/UpdateHorarioUseCase.java
@@ -5,5 +5,5 @@ import com.example.finca_hexagonal.domain.models.Horario;
 import java.util.Optional;
 
 public interface UpdateHorarioUseCase {
-    Optional<Horario> updateHorario(int horarioId, Horario horario);
+    Optional<Horario> updateHorario(Long horarioId, Horario horario);
 }
